rename metakg.index to metakg.api
